login api access token and refreshtoken issue fix

+++ internal/model/user_model.go
@@ -15,7 +15,7 @@ type (
 
 	RefreshTokenModel struct {
 		ID        int
-		UserID    int
+		UserID    int64
 		Token     string
 		ExpiresAt time.Time
 		CreatedAt time.Time
